+++ src/StorageBase/Action/PortalExtension/Deactivate/PortalExtensionDeactivatePayload.php
@@ -4,51 +4,8 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Deactivate;
 
-use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
-use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
-use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
-use Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionType;
-use Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionTypeCollection;
-use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
+use Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Contract\PortalExtensionActiveChangePayloadContract;
 
-final class PortalExtensionDeactivatePayload implements AttachmentAwareInterface
+final class PortalExtensionDeactivatePayload extends PortalExtensionActiveChangePayloadContract
 {
-    use AttachmentAwareTrait;
-
-    private PortalNodeKeyInterface $portalNodeKey;
-
-    private PortalExtensionTypeCollection $extensions;
-
-    public function __construct(PortalNodeKeyInterface $portalNodeKey)
-    {
-        $this->attachments = new AttachmentCollection();
-        $this->portalNodeKey = $portalNodeKey;
-        $this->extensions = new PortalExtensionTypeCollection();
-    }
-
-    public function getPortalNodeKey(): PortalNodeKeyInterface
-    {
-        return $this->portalNodeKey;
-    }
-
-    public function addExtension(PortalExtensionType $portalExtensionType): void
-    {
-        if ($this->extensions->contains($portalExtensionType)) {
-            return;
-        }
-
-        $this->extensions->push([$portalExtensionType]);
-    }
-
-    public function removeExtension(PortalExtensionType $portalExtensionType): void
-    {
-        $this->extensions = $this->extensions->filter(
-            static fn (PortalExtensionType $item): bool => !$item->equals($portalExtensionType)
-        );
-    }
-
-    public function getExtensions(): PortalExtensionTypeCollection
-    {
-        return new PortalExtensionTypeCollection($this->extensions);
-    }
 }